backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.5
mangum==0.19.0
pymongo==4.9.2
# bson==0.5.10
motor==3.6.1
beanie==1.29.0
python-dotenv==1.0.1
langchain==0.3.14
langchain-community==0.3.14
langgraph==0.2.64
langchain-groq==0.2.3
python-multipart==0.0.20
pinecone==5.4.2
pypdf==5.1.0
python-docx==1.1.2
pillow==11.1.0
pytesseract==0.3.13
boto3==1.36.2
pymupdf4llm==0.0.17
arize-phoenix==7.9.4
arize-phoenix-otel==0.7.1
openinference-instrumentation-langchain==0.1.30
tavily-python==0.5.0
mediawikiapi==1.2
wikipedia==1.4.0
arxiv==2.1.3
unstructured==0.16.20
poppler-utils==0.1.0
